Created second story instance
This is primarily for proof-of-concept. Instantiating two stories locally is really bad practice, but it gets the point across in terms of having separate stories functionally coexist. I've dropped myself a bunch of TODOs for when I go back to refactor; thankfully each component is designed to scale well enough to where I won't have to double/triple/quadruple the code each time I want to add a new story.